fix: sadly remove some dev commands from docker-compose
Docker compose unfortunately does not work well for our frontend.
The file lookup through bindmounts for the many files in a frontend
application make it extremely sluggish and in many cases result in docker
container exit status's that are unpredictable.
Yarn 2 might make it possible to use docker-compose
for development again in a more efficient way,
due to the fact it doesn't use symlinks for yarn workspaces.

Signed-off-by: Mat Appelman <mat@opstrace.com>