fix: fix dangerous & unnecessary mapping of types
Having a mapping the way it was, meant it is too easy to mistakenly rename the key,
resulting in runtime errors because the deserialize method doesn't actually return
the correct type. The compiler cannot type check here because we are marshalling
JSON to/from a string, so we need to be as cautious as possible.

Signed-off-by: Mat Appelman <mat@opstrace.com>