cli: type-tweak smErrorLastResort()
TSC error message:

   Types of parameters 'detail' and 'detail' are incompatible.
     Type 'unknown' is not assignable to type '{ sagaStack: string; }'.

 55   await destroyCluster(util.smErrorLastResort);

redux-saga calls the function with the `unknown` type,
for the `detail` arg. Define this function with `any`
type for the `detail` arg. That is "compatible" with
`unknown`. If we were to keep `unknown` then it's a
little difficult to actually get to the `sagaStack`
string property (at least for me).

Compilation succeeded locally. Pushing!

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>