fix: bring up loki and cortex statefulset services in order
The loki and cortex services need to be brough up in order otherwise they will
try to join the membership ring simultaneously.

From Kubernetes StatefulSet documentation:

For a StatefulSet with N replicas, when Pods are being deployed, they are
created sequentially, in order from {0..N-1}.  When Pods are being deleted,
they are terminated in reverse order, from {N-1..0}.  Before a scaling
operation is applied to a Pod, all of its predecessors must be Running and
Ready.  Before a Pod is terminated, all of its successors must be completely
shutdown.

https://kubernetes.io/docs/concepts/workloads/controllers/statefulset

Signed-off-by: Simão Reis <sreis@opstrace.com>