chore: start using typed-redux-saga for #235
See
https://github.com/opstrace/opstrace/issues/235


Not just the import changed, but also:

  yield -> yield*

"The yield* expression is used to delegate to another
generator or iterable object."

`yield * iterable` is basically
`for (const value of iterable) yield value`

With this change, tsc showed an error:

Type 'unknown' is not assignable to type 'string[]'.ts(2322)

Which was the goal. Nice.

In this patch: make redux thingies an explicit
dependency of the cli and the controller, not of
the main workspace. Also regenerate yarn.lock.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>