go: proxy_test: address url confusion
This was misleading me:

    httptest.NewRequest("GET", u.String(), nil)

the URL being used here is the URL to the
proxy upstream/backend. And the HTTP request
being constructed here is supposed to go to
the frontend side of the proxy.

Any URL works here because we test the proxy
handlers _directly_, bypassing a routing
layer.

So, for now I set this to http://localhost and
add a code comment.

Also: use assert.Equal(), and add a bit of commentary
Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>

wip: squash into previous

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>