controller: Add config API pod to application deployment (#355)
This deploys the `config` service from `go/cmd/config/`. It serves endpoints for adding/updating/removing credentials and exporters in the GraphQL service. It is accessible at e.g. `config.cluster-name.opstrace.io`. The endpoints are currently documented in `go/cmd/config/README.md`, planning on writing a nice user-facing guide once the controller sync support is complete.

The controller deploys a single common `config` ingress/service/pod into the `application` namespace, rather than the per-tenant namespaces, so that it can use the existing Hasura access secrets in the `application` namespace. This felt like a better tradeoff than running instances in each of the tenant namespaces, which would have meant needing to add a copy of the Hasura secret into the per-tenant namespaces as well. So, when the config service receives a request over HTTP, it extracts the tenant name from the (required and validated) bearer token, rather than being "locked" to a single tenant. See also #351.

As of this PR, we support writing/updating/deleting credentials and exporter configs in GraphQL via `curl`. The interface itself has pretty good validation and error handling for the configs, and should be pretty much ready to go. However, there isn't anything that actually consumes the configs yet. Next steps:
- Updating the controller itself to automatically sync GraphQL credentials/exporters into K8s tenant namespaces as Secrets and Deployments.
- Once things are known-working, writing a user guide for the docs describing how to configure and deploy exporters (and their prerequisite credentials) in various environments, using `curl` commands to the config service.

Signed-off-by: Nick Parker <nick@opstrace.com>