looker: ts fragment pusher: sleep between 5xx retries
Fixes an issue in
TimeseriesFragmentPushMessage.postWithRetryOrError()
where two no waiting time was spent between POST
HTTP requests that were responded to with 5xx type
responses (that really was an oversight, a missing
sleep() statement).

This is to address #480.

This instability newly showed itself as of
the 429 response rewriting implemented via
#464.

429 responses were handled by design (with a small
wait step inbetween attempts).

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>