cortex: set compactor_blocks_retention_period
Let the compactor delete old blocks. Note that
the definition is super important:
"Delete blocks containing samples older than the"
... that is, the block might container newer
samples, too.

Don't overthink this yet, add a buffer. Also,
it's unclear if this is doing anything at all or
of this is superseded by the bucket lifecycle-based
deletion method.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>