test-remote: Add test for azure exporter (#556) (#590)
* test-remote: Add test for azure exporter

The Azure exporter fails to start when provided with bogus credentials, so we just check its logs to validate that it was deployed and that it failed to query Azure as expected. In comparison other exporters expose a metric with a nonzero error count.

Signed-off-by: Nick Parker <nick@opstrace.com>

* test-remote: Extend timeout for exporter log, system-tenant seems to take longer

The exporter log for system-tenant seems to take longer than for default-tenant, maybe its configured differently? For example in a manual run I got 62s:
```
2021-04-26T23:22:57.986Z info: got expected result in query 60 after 62.77 s
```
Then in another run I got 37s:
```
2021-04-26T23:29:58.337Z info: got expected result in query 36 after 37.65 s
```

Signed-off-by: Nick Parker <nick@opstrace.com>