chore: memcached: mem requests with more headroom
Depending on exact unit definition and
measurement technique, the k8s memory
request for memcached containers needs to
be tuned to the -m XX command line option
of memcached, with headroom applied.
Headroom is necessary for potentially
different definition of units, and more
importantly for the ambiguity of how the
-m XX arg relates to _actual_ mem consumption.

Note that -m XX is documented with
"Set the amount of memory allocated to
memcached for object storage."

Nothing in this statement indicates that
when setting e.g. -x 1024 that the memory
consumption as measured/seen by k8s is
limited to pretty much exactly ~1024 MB.
A memcached process consumes memory for
other things than the object storage alone.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>