looker: logEveryNcalc()
A more generic solution for throttling
log messages around similar events.

To stay with the code example, if unthrottled
log messages would be 400000 then the goal is
to log every 10000th log message so that after
all about 40 log messages pop up. That is, this
solution keeps the number of actual log messages
emitted on the order of 10^1.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>