test-remote: query loki labels through grafana
Query the system log labels in Loki
through the Grafana system, i.e.
through the Grafana datasource proxy
and (as of today) also through the
datasource splitter reverse proxy
which we did introduce between the
Grafana datasource proxy and the
actual data source.

This uses the UI's (browser's) authentication
state.

This is towards bisecting #908

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>