controller: datasource splitter ngx conf work
If the goal is to rewrite parts of the URL and to
use the DNS caching workaroundThe proxy_pass URL
rewriting is _not_ performed.

The generic solution is to extract the relevant portion
in the incoming URL with regex in the location directive
and to re-use that in the proxy_pass statement.

when one only wants to carry over query args,
one can use $is_args$query_string

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>