looker: remove complexity from image build (yarn.lock)
This was a PITA in the last 60 mins -- not sure why,
but I think the idea of using a massive yarn.lock
file and then installing only a subset is kind of
flawed. yarn did a huge removal on yarn.lock and
that failed the command. Want to keep the
independence, so no global yarn.lock for now.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>