looker: Shift label randomness into invocation_id, allow configuration
Randomness is provided via `invocation_id`, and then uniqueness across cycles and concurrent_streams is added on top via the indexes. End result:
- Metrics:
    - metricName: `<invocation_id>-<cycle#>`
    - uniqueName: `<invocation_id>-<cycle#>-<stream#>`
- Logs:
    - uniqueName: `<invocation_id>-<cycle#>-<stream#>`

With the randomness moved into `invocation_id` (and with it exposed in the args), it can be overridden manually, allowing e.g. series counts to remain stable across looker runs.

Signed-off-by: Nick Parker <nick@opstrace.com>