ci: remove "config" prefix
The config API is part of the go project,
so maybe we want to prefix those with "go: ".

Or we want to rename the config API because
it doesn't seem right, yet.

Is it part of the tenant API?

`config: ` just seems to be too misleading
when seen in a commit message.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>