ci: remove feat/fix commit prefixes
As also explained in the commitlint config
file and discussed an some github issues,
these make sense _only_ when talking about
user-facing aspects, and in a world where
semantic versioning is used. In the majority
of the cases I have seen these being
'misused' in our case. For example, a "fix"
in a test suite is never a user-facing fix.
Let's try to do without, for now.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>