ci: preamble: regen docker-image.json early, more commits
Start this in background. Then use the new
WRITE_NEW_DOCKER_IMAGES_JSON_FILE_HERE_ABSPATH
mechanism to wait for the newly generated
file to be written. Only then proceed
with building the controller image.

It is critically important to build the controller
image with the _updated_ docker-image.json.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>

ci: preamble: controller img build in separate directory

Now that we start building things concurrently
in the preamble it is also important to isolate things
Execute the controller image build in its own
directory.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>

ci: preamble: pull up linting

Also fix a -e/-x confusion (important!)

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>

ci: preamble: wait for img builds towards the end

These two potentially take a 'long' time
and therefore it makes most sense to
inspect the result at the end of the preamble,
to do as many 'other' things as possible
in the meantime.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>