cli: gcp subnet destroy: require region match
There may be an asymmetry between the
"region to destroy in" and the region of
the subnetwork found during the lookup.

The network does not even have "opstrace"
in its name so let's not blindly delete
in this case.

This needs to be improved conceptually
by doing proper region detection upon
destroy (or requesting user input), see

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>
#1290