looker: walltime coupling: dynamic forward-leap
This is a significant behavioral change for
the walltime coupling method.

When fallen behind into the past beyond
the maximum lag criterion, do not leap
forward by a fixed amount -- but dynamically
leap forward to the right end of the
'good regime', i.e. shortly before the minimum
lag criterion.

This for example helps when we've fallen behind by a
_lot_ (e.g., through putting the machine to sleep).

Also, the dynamics are very much determined by
where in said 'good regime' we actually land
after leaping forward. When landing towards the
right end, we maximize the time it takes before
the next forward leap is required (assuming that
the synthetic time keeps falling behind at a
certain rate).

This conceptually removed the forward leap
distance as an interesting input parameter:
if smaller leaps are the goal, then tuning
maxLagSeconds and minLagSeconds so that
they are closer to each other suffices.

Hence, this commit also

- simplifies building the interval based on
  which a time series start time is chosen
  (and does not expose these boundaries
  to the user anymore

- removes now outdated parameters from
  the CLI

Will expose the minLax and maxLag concepts
via CLI in a follow-up patch.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>